+++ backend/pcloud-api/src/main/java/com/api/exhibition/presentation/ExhibitionController.java
@@ -7,8 +7,8 @@ import com.api.exhibition.application.dto.ExhibitionUpdateRequest;
 import com.api.exhibition.presentation.dto.ExhibitionLikedStatusResponse;
 import com.domain.annotation.AuthMember;
 import com.domain.annotation.AuthMembers;
-import com.domain.show.exhibition.infrastructure.dto.ExhibitionSimpleResponse;
-import com.domain.show.exhibition.infrastructure.dto.ExhibitionSpecificResponse;
+import com.domain.show.exhibition.domain.dto.ExhibitionSimpleResponse;
+import com.domain.show.exhibition.domain.dto.ExhibitionSpecificResponse;
 import java.net.URI;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
