+++ backend/pcloud-api/src/test/java/com/api/exhibition/presentation/ExhibitionControllerWebMvcTest.java
@@ -3,8 +3,8 @@ package com.api.exhibition.presentation;
 import com.api.exhibition.application.dto.ExhibitionCreateRequest;
 import com.api.exhibition.application.dto.ExhibitionUpdateRequest;
 import com.api.helper.MockBeanInjection;
-import com.domain.show.exhibition.infrastructure.dto.ExhibitionSimpleResponse;
-import com.domain.show.exhibition.infrastructure.dto.ExhibitionSpecificResponse;
+import com.domain.show.exhibition.domain.dto.ExhibitionSimpleResponse;
+import com.domain.show.exhibition.domain.dto.ExhibitionSpecificResponse;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import java.util.List;
 import java.util.Optional;
