+++ backend/pcloud-domain/src/main/java/com/domain/show/exhibition/infrastructure/ExhibitionRepositoryImpl.java
@@ -3,8 +3,8 @@ package com.domain.show.exhibition.infrastructure;
 import com.domain.show.exhibition.domain.Exhibition;
 import com.domain.show.exhibition.domain.ExhibitionRepository;
 import com.domain.show.exhibition.domain.LikedExhibition;
-import com.domain.show.exhibition.infrastructure.dto.ExhibitionSimpleResponse;
-import com.domain.show.exhibition.infrastructure.dto.ExhibitionSpecificResponse;
+import com.domain.show.exhibition.domain.dto.ExhibitionSimpleResponse;
+import com.domain.show.exhibition.domain.dto.ExhibitionSpecificResponse;
 import java.util.List;
 import java.util.Optional;
 import lombok.RequiredArgsConstructor;
