+++ backend/pcloud-domain/src/testFixtures/java/show/exhibition/infrasturcture/ExhibitionFakeRepository.java
@@ -3,8 +3,8 @@ package show.exhibition.infrasturcture;
 import com.domain.show.exhibition.domain.Exhibition;
 import com.domain.show.exhibition.domain.ExhibitionRepository;
 import com.domain.show.exhibition.domain.LikedExhibition;
-import com.domain.show.exhibition.infrastructure.dto.ExhibitionSimpleResponse;
-import com.domain.show.exhibition.infrastructure.dto.ExhibitionSpecificResponse;
+import com.domain.show.exhibition.domain.dto.ExhibitionSimpleResponse;
+import com.domain.show.exhibition.domain.dto.ExhibitionSpecificResponse;
 import java.util.Comparator;
 import java.util.HashMap;
 import java.util.List;
